feat: 마이페이지에서 업데이트 버튼 클릭 시 재로그인 로직 추가
